refactor(core): 重构代码结构，移动多个模块至 core 文件夹
- 将 ui 目录下的多个模块移动至 ui/core 目录，调整相关导入路径以适配新结构
- 将 common 目录下部分模块重命名移动至 common/runtime 和 common/system，更新导入引用
- 更新测试文件中相关导入路径，确保测试代码正常执行
- 新增 session/utils.ts，抽取并集中管理 getCompactPromptTokenThreshold 与 getTotalTokens 方法
- 调整部分导入钩子名称由 prompt 变更为 hooks，提升代码组织一致性
- 修正部分日志及进程管理模块的导入路径到对应的 logging 和 system 子目录
- 更新 UI 组件和工具相关代码以适应新的模块路径和结构变化
