+++ src/common/logging/error-logger.ts
