+++ src/common/system/bash-timeout.ts
