+++ src/common/updateCheck.ts
@@ -6,7 +6,7 @@ import * as path from "path";
 import { render, type Instance } from "ink";
 import chalk from "chalk";
 import { UpdatePrompt, type UpdatePromptChoice } from "../ui";
-import { killProcessTree } from "./process-tree";
+import { killProcessTree } from "./system/process-tree";
 
 export type PackageInfo = {
   name: string;
