+++ src/session-types.ts
@@ -1,7 +1,6 @@
 import type { McpServerConfig, PermissionScope, PermissionSettings } from "./settings";
 import type { AskPermissionRequest, MessageToolPermission, UserToolPermission } from "./common/permissions";
 import type { CreateOpenAIClient } from "./tools/executor";
-import { DEEPSEEK_V4_MODELS } from "./common/model-capabilities";
 
 export type SessionStatus =
   | "failed"
@@ -139,24 +138,3 @@ export type LlmStreamProgress = {
   formattedTokens: string;
   phase: "start" | "update" | "end";
 };
-
-const DEFAULT_COMPACT_PROMPT_TOKEN_THRESHOLD = 128 * 1024;
-const DEEPSEEK_V4_COMPACT_PROMPT_TOKEN_THRESHOLD = 512 * 1024;
-
-export function getCompactPromptTokenThreshold(model: string): number {
-  return DEEPSEEK_V4_MODELS.has(model)
-    ? DEEPSEEK_V4_COMPACT_PROMPT_TOKEN_THRESHOLD
-    : DEFAULT_COMPACT_PROMPT_TOKEN_THRESHOLD;
-}
-
-function isUsageRecord(value: unknown): value is Record<string, unknown> {
-  return value !== null && typeof value === "object" && !Array.isArray(value);
-}
-
-export function getTotalTokens(usage: ModelUsage | null | undefined): number {
-  if (!isUsageRecord(usage)) {
-    return 0;
-  }
-  const totalTokens = (usage as Record<string, unknown>).total_tokens;
-  return typeof totalTokens === "number" ? totalTokens : 0;
-}
