+++ src/tests/clipboard.test.ts
@@ -5,7 +5,7 @@ import * as os from "os";
 import * as path from "path";
 
 // eslint-disable-next-line @typescript-eslint/consistent-type-imports
-type ClipboardModule = typeof import("../ui/clipboard");
+type ClipboardModule = typeof import("../ui/core/clipboard");
 
 const ORIGINAL_PATH = process.env.PATH;
 const ORIGINAL_PLATFORM = process.platform;
@@ -30,7 +30,7 @@ function withPlatform<T>(platform: NodeJS.Platform, fn: () => T): T {
 
 test("readClipboardImage returns null when no clipboard helpers are installed", async () => {
   // Reload module so it picks up the patched PATH at spawn time.
-  const moduleUrl = new URL(`../ui/clipboard.ts?t=${Date.now()}`, import.meta.url).href;
+  const moduleUrl = new URL(`../ui/core/clipboard.ts?t=${Date.now()}`, import.meta.url).href;
   const { readClipboardImage } = (await import(moduleUrl)) as ClipboardModule;
   const result = withCleanPath(() => readClipboardImage());
   assert.equal(result, null);
@@ -63,7 +63,7 @@ test(
         { mode: 0o755 }
       );
 
-      const moduleUrl = new URL(`../ui/clipboard.ts?t=${Date.now()}`, import.meta.url).href;
+      const moduleUrl = new URL(`../ui/core/clipboard.ts?t=${Date.now()}`, import.meta.url).href;
       const { readClipboardImage } = (await import(moduleUrl)) as ClipboardModule;
 
       process.env.PATH = binDir;
