+++ src/tools/write-handler.ts
@@ -9,8 +9,14 @@ import {
   readTextFileWithMetadata,
   writeTextFile,
 } from "../common/file-utils";
-import { executeValidatedTool } from "../common/runtime";
-import { getFileState, isAbsoluteFilePath, isFullFileView, normalizeFilePath, recordFileState } from "../common/state";
+import { executeValidatedTool } from "../common/runtime/runtime";
+import {
+  getFileState,
+  isAbsoluteFilePath,
+  isFullFileView,
+  normalizeFilePath,
+  recordFileState,
+} from "../common/runtime/state";
 
 const writeSchema = z.strictObject({
   file_path: z.string().min(1, "file_path is required."),
