+++ src/ui/components/FileMentionMenu/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Box, Text } from "ink";
 import { useInput } from "ink";
 import DropdownMenu from "../DropdownMenu";
-import type { FileMentionItem, FileMentionToken } from "../../fileMentions";
+import type { FileMentionItem, FileMentionToken } from "../../core/fileMentions";
 
 type Props = {
   open: boolean;
