+++ src/ui/core/promptBuffer.ts
