+++ src/ui/hooks/useTerminalInput.ts
